local: Fix NaN values in sun set times
In spring, the next sun set can be in more than 24 hours which
astroplan can not handle if only the next sun set from the current
time (which is also at a sun set time) is calculated. Therefore, check
another time one hour later to get a next sun set which is within 24
hours.